+++ database/migrations/0001_01_01_000001_create_cache_table.php
@@ -11,6 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (config('cache.default') !== 'database') return;
+
         Schema::create('cache', function (Blueprint $table) {
             $table->string('key')->primary();
             $table->mediumText('value');
@@ -29,6 +31,8 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (config('cache.default') !== 'database') return;
+
         Schema::dropIfExists('cache');
         Schema::dropIfExists('cache_locks');
     }
